Initial revision
FossilOrigin-Name: 093bd6983aa31bf6cf8b2566dc9bf0ccc16760c6be92d3afa0d48d455ae7de9e

+++ tty/termcap/ttyio.c
@@ -21,7 +21,6 @@
 #endif
 
 #include	<errno.h>
-#include	<unistd.h>
 #include	<sys/ioctl.h>
 #ifdef sun
 #include	<sys/filio.h>
@@ -176,7 +175,7 @@ ttclose (void)
 int
 ttstat (void)
 {
-  int n;
+  long n;
 
   if (ioctl (0, FIONREAD, &n) < 0)
     return (FALSE);
